Temat_9 Dodanie wielowątkowości testów
Atrybut thread-count – określający, ile wątków maksymalnie ma być uruchomionych naraz
Atrybut parallel – określający jak TestNG powinien definiować jeden wątek. Gdzie jednym wątkiem może być, w zależności od wartości pola na przykład tag test, tag classes, tag methods, czy po prostu zapisana na twardo ilość instancji (instances).
