+++ src/test/java/JavaStartSeleniumRozdzial_9/driver/DriverUtils.java
@@ -1,6 +1,8 @@
 package JavaStartSeleniumRozdzial_9.driver;
 
 
+import JavaStartSeleniumRozdzial_9.driver.manager.DriverManager;
+
 public class DriverUtils {
 
     public static void setInitialConfiguration() {
