+++ src/test/java/JavaStartSeleniumRozdzial_9/page/objects/CheckoutPage.java
@@ -1,7 +1,7 @@
 package JavaStartSeleniumRozdzial_9.page.objects;
 
 
-import JavaStartSeleniumRozdzial_9.driver.DriverManager;
+import JavaStartSeleniumRozdzial_9.driver.manager.DriverManager;
 import org.openqa.selenium.support.PageFactory;
 
 public class CheckoutPage {
