+++ src/test/java/JavaStartSeleniumRozdzial_9/tests/TestBase.java
@@ -2,7 +2,7 @@ package JavaStartSeleniumRozdzial_9.tests;
 
 import JavaStartSeleniumRozdzial_9.configuration.ConfigurationProperties;
 import JavaStartSeleniumRozdzial_9.configuration.PropertiesLoader;
-import JavaStartSeleniumRozdzial_9.driver.DriverManager;
+import JavaStartSeleniumRozdzial_9.driver.manager.DriverManager;
 import JavaStartSeleniumRozdzial_9.driver.DriverUtils;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeClass;
