+++ src/test/resources/all_tests_suite.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd" >
-<suite name="All Pet Store Tests">
+<suite name="All Pet Store Tests" parallel="tests" thread-count="3">
 
     <test name="Negative Login Tests">
         <classes>
@@ -11,6 +11,12 @@
     <test name="Positive Login Tests">
         <classes>
             <class name="JavaStartSeleniumRozdzial_9.tests.PositiveLoginTests"/>
+<!--            <class name="JavaStartSeleniumRozdzial_9.tests.ShoppingCartTest"/>-->
+        </classes>
+    </test>
+    
+    <test name="Shopping Cart Tests">
+        <classes>
             <class name="JavaStartSeleniumRozdzial_9.tests.ShoppingCartTest"/>
         </classes>
     </test>
